Add utils module for the common types
