+++ domains/client/domain-executor/src/utils.rs
@@ -1,0 +1,35 @@
+use sc_consensus::ForkChoiceStrategy;
+use sp_consensus_slots::Slot;
+use sp_runtime::traits::{Block as BlockT, NumberFor};
+use std::fmt::Debug;
+use subspace_core_primitives::Blake2b256Hash;
+
+/// Data required to produce bundles on executor node.
+#[derive(PartialEq, Clone, Debug)]
+pub(super) struct ExecutorSlotInfo {
+    /// Slot
+    pub(super) slot: Slot,
+    /// Global challenge
+    pub(super) global_challenge: Blake2b256Hash,
+}
+
+/// An event telling the `Overseer` on the particular block
+/// that has been imported or finalized.
+///
+/// This structure exists solely for the purposes of decoupling
+/// `Overseer` code from the client code and the necessity to call
+/// `HeaderBackend::block_number_from_id()`.
+#[derive(Debug, Clone)]
+pub struct BlockInfo<Block>
+where
+    Block: BlockT,
+{
+    /// hash of the block.
+    pub hash: Block::Hash,
+    /// hash of the parent block.
+    pub parent_hash: Block::Hash,
+    /// block's number.
+    pub number: NumberFor<Block>,
+    /// Fork choice of the block.
+    pub fork_choice: ForkChoiceStrategy,
+}
